requirements.txt
playwright==1.44
python-telegram-bot==21.1
google-cloud-firestore==2.15.0
pydantic==2.8.0
python-dotenv==1.0.1
requests
pytest==8.2.0

Flask==3.0.0
